requirements.txt
PyMuPDF
Pillow
pyinstaller
